requirements.txt
textual>=0.47.0
rich>=13.7.0
ollama>=0.2.0
openai>=1.12.0
pydantic>=2.5.0
typer>=0.9.0
questionary>=2.0.1
pygments>=2.17.0
pyfiglet>=1.0.2
python-dotenv>=1.0.0
httpx>=0.26.0
